Move panic handler into define_zome! macro to only add it to the zome project using the macro (to fix Windows CI)
