+++ hdk-rust/src/lib.rs
@@ -46,5 +46,3 @@ pub mod meta;
 
 pub use crate::api::*;
 pub use holochain_core_types::validation::*;
-
-mod panic_handler;
